+++ CHANGES.md
@@ -1,5 +1,9 @@
 #### [unreleased]
 
+#### 10.8.0 / 2022-04-03
+* use `move_dir()` and `is_virtualbox()` from [#51875](https://core.trac.wordpress.org/ticket/51857) [PR #2225](https://github.com/WordPress/wordpress-develop/pull/2225/files)
+* revert fix directory rename for single file plugin update
+
 #### 10.7.2 / 2022-03-31
 * fix directory rename for single file plugin update
 * revert GitHub_API release asset URL to return to redirect URL